Se ha avanzado en el driver de dispositivos de bloques DB.C. El sistema de drivers se ha mejorado de manera que ante una llamada (open, read, write, ...) el segmento de datos del drive se obtiene a partir del campo segDatosSR de la cabecera, siendo innecesario declarar una funcion segDatos. Ademas la llamada al sistema de drivers esperarDesinstalacion se ha ampliado para permitir eliminar partes de codigo, datos y pila innecesarios del proceso driver instalado, e indicar la funcion finishDriver que queremos que se ejecute tras la desinstalacion. La idea es situar el codigo necesario al principio del segmento de datos hasta una funcion finDriver y mover hasta esa direccion los datos (inicializados, DATA) necesarios, añadiendo a continuacion una pila mínima conteniendo solo la trama (DS, ES, ... , IP, CS, Flags) que retorne a la funcion finishDriver. Actualmente esto funciona perfectamente para los drivers RELOJ y RETARDO, reduciendose significativamente la memoria que utilizan tras su instalacion.

+++ PRACT0/SO1/MANEJ_01.C
@@ -79,8 +79,8 @@ void so1_manejador_01 ( void ) {                       /* ah = 1 ; int SO1 */
   modoAp_t modoAp ;
   pointer_t dir ;
   word_t nbytes ;
-  long int posNueva ;
-  long int pos ;
+  long posNueva ;
+  long pos ;
   int whence ;
   word_t cmd ;                                                    /* fcntl */
   word_t arg ;
@@ -191,17 +191,17 @@ void so1_manejador_01 ( void ) {                       /* ah = 1 ; int SO1 */
 
     case 0x06 :                                                 /* 0x0106 */
       df = tramaProceso->BX ;                                    /* lseek */
-      posNueva = (long int)-1 ;
+      posNueva = -1L ;
       if ((0 <= df) && (df < dfMax)) {
         dfs = descProceso[indProcesoActual].tfa[df].dfs ;
         if ((0 <= dfs) && (dfs < dfsMax)) {
           rindx = descFichero[dfs].rindx ;
           if ((0 <= rindx) && (rindx < maxRecursos)) {
-            pos = (long int)pointer(tramaProceso->ES, tramaProceso->DX) ;
+            pos = (long)pointer(tramaProceso->ES, tramaProceso->DX) ;
             whence = tramaProceso->CX ;
             if ((SEEK_SET <= whence) && (whence <= SEEK_END)) {
               posNueva = descRecurso[rindx].lseek(dfs, pos, whence) ;
-              if (posNueva >= (long int)0)
+              if (posNueva >= (long)0)
                 descProceso[indProcesoActual].tfa[df].pos = posNueva ;
             }
           }
